perf(Front/Scatter): Improve reduced features and trajectories tracing
